feat(nexrad): cut adapter / attrs / mod over to internal decode_volume
Phase 7b: replace `nexrad`/`nexrad-data`/`nexrad-decode`/`nexrad-model`
runtime calls with `crate::backends::nexrad::decode::decode_volume`
across the backend layer. Net diff is -169 lines (more compact
because the upstream MSG_2 walk + waveform/channel-config enum
round-trips collapse).

Changes:

* `mod.rs`: replaced `nexrad::data::volume::File::new(data).scan()`
  + the second-pass `first_msg2()` walk with a single
  `decode_volume(&data)` call. The new `Scan` already carries
  `rda_status` so MSG_2 doesn't need a separate decompress + walk.
  Net: -75 lines.

* `adapter.rs`: switched to `decode::model::Scan` / `Sweep` /
  `Radial` / `Site` and `decode::messages::msg5::ElevationCut`
  + `decode::products::{Product, DataMoment, MomentValue,
  CfpMomentValue}`. Field-access where upstream used getters
  (`scan.sweeps` not `scan.sweeps()`, `r.azimuth_angle_degrees`
  not `r.azimuth_angle_degrees()`). `convert_scan` no longer
  needs an `msg2` parameter — it reads `scan.rda_status` instead.

* `attrs.rs`: dropped the `WaveformType`/`ChannelConfiguration`/
  `PulseWidth` enum round-trips (`waveform_to_raw` /
  `channel_config_to_raw` / `pulse_width_str` helpers) — our raw
  u8 fields feed xradar's lookup tables directly. Test fixtures
  rewritten to struct-literal form with a `zero_cut_with` helper.

* `mapping.rs`: 2-line import swap to use our `Product`.

* `decode/products.rs`: `DataMoment` trait now returns `f64` so
  `MomentGeometry` (also f64) gets a clean `m.first_gate_range_km()`
  call without per-callsite casts.

Verification:

* 194 unit tests pass, 5 fixture-gated tests pass against live
  KLOT (12 sweeps / 7200 rays) and KILX (13 sweeps / 6840 rays)
* 2 structural parity tests pass against `danielway/nexrad`
* `cargo fmt`, `cargo clippy --workspace --all-targets -- -D warnings`
  clean

6-skill audit (post-cut-over diff):
  - /simplify: HIGH/MED clean. LOW (skipped): test-fixture
    `OwnedMoment` literal repeats 3×; per-variant differences
    make extraction marginal.
  - /rust-best-practices: borrow-not-clone, no unwrap in prod,
    pub(super) minimal, exhaustive matches — clean.
  - /rust-engineer, /rust-patterns: idiomatic field access vs
    method calls; iterator chains preserved.
  - /rust-testing: 9 attrs tests + 12 adapter tests preserved
    behaviors (xradar parity strings, sweep counts, fixed-angle
    preference); zero-cut-with fixture reduces boilerplate.
  - /rust-async-patterns: N/A.

`nexrad`/`nexrad-data`/`nexrad-decode`/`nexrad-model` are still
runtime deps until Phase 7c drops them from `Cargo.toml`.

+++ radish/src/backends/nexrad/decode/products.rs
@@ -12,19 +12,21 @@ use super::model::{OwnedCfp, OwnedMoment, Radial};
 
 /// Common geometry for any per-radial moment block. Used by the
 /// adapter's `probe_geometry` to size the (rays × gates) array
-/// without decoding gate values.
+/// without decoding gate values. f64 return matches
+/// `backends::common::MomentGeometry` so the adapter doesn't need
+/// per-call casts.
 pub(crate) trait DataMoment {
-    fn first_gate_range_km(&self) -> f32;
-    fn gate_interval_km(&self) -> f32;
+    fn first_gate_range_km(&self) -> f64;
+    fn gate_interval_km(&self) -> f64;
     fn gate_count(&self) -> u16;
 }
 
 impl DataMoment for OwnedMoment {
-    fn first_gate_range_km(&self) -> f32 {
-        self.descriptor.range_to_first_gate_km
+    fn first_gate_range_km(&self) -> f64 {
+        f64::from(self.descriptor.range_to_first_gate_km)
     }
-    fn gate_interval_km(&self) -> f32 {
-        self.descriptor.gate_interval_km
+    fn gate_interval_km(&self) -> f64 {
+        f64::from(self.descriptor.gate_interval_km)
     }
     fn gate_count(&self) -> u16 {
         self.descriptor.gate_count
@@ -32,11 +34,11 @@ impl DataMoment for OwnedMoment {
 }
 
 impl DataMoment for OwnedCfp {
-    fn first_gate_range_km(&self) -> f32 {
-        self.descriptor.range_to_first_gate_km
+    fn first_gate_range_km(&self) -> f64 {
+        f64::from(self.descriptor.range_to_first_gate_km)
     }
-    fn gate_interval_km(&self) -> f32 {
-        self.descriptor.gate_interval_km
+    fn gate_interval_km(&self) -> f64 {
+        f64::from(self.descriptor.gate_interval_km)
     }
     fn gate_count(&self) -> u16 {
         self.descriptor.gate_count
